refactor(tools): move regression code into separate cpp files
