+++ tools/acl_compressor/sources/acl_compressor.cpp
@@ -41,22 +41,13 @@
 #endif
 
 #include "acl/core/ansi_allocator.h"
-#include "acl/core/compressed_database.h"
 #include "acl/core/floating_point_exceptions.h"
-#include "acl/core/iallocator.h"
-#include "acl/core/range_reduction_types.h"
 #include "acl/core/string.h"
 #include "acl/core/impl/debug_track_writer.h"
 #include "acl/compression/compress.h"
-#include "acl/compression/convert.h"
-#include "acl/compression/transform_error_metrics.h"
 #include "acl/compression/transform_pose_utils.h"	// Just to test compilation
 #include "acl/compression/impl/write_decompression_stats.h"
-#include "acl/compression/track_error.h"
 #include "acl/decompression/decompress.h"
-#include "acl/decompression/database/database.h"
-#include "acl/decompression/database/null_database_streamer.h"	// Just to test compilation
-#include "acl/decompression/database/impl/debug_database_streamer.h"
 #include "acl/io/clip_reader.h"
 
 #include <cstring>
@@ -445,1025 +436,16 @@ static bool parse_options(int argc, char** argv, Options& options)
 	return true;
 }
 
-struct debug_transform_decompression_settings_with_db : public acl::debug_transform_decompression_settings
-{
-	using database_settings_type = acl::debug_database_settings;
-};
-
 #if defined(ACL_USE_SJSON)
 #if defined(ACL_HAS_ASSERT_CHECKS)
-static void validate_accuracy(iallocator& allocator, const track_array_qvvf& raw_tracks, const track_array_qvvf& additive_base_tracks, const itransform_error_metric& error_metric, const compressed_tracks& compressed_tracks_, double regression_error_threshold)
-{
-	using namespace acl_impl;
-
-	// Disable floating point exceptions since decompression assumes it
-	scope_disable_fp_exceptions fp_off;
-
-	acl::decompression_context<debug_transform_decompression_settings_with_db> context;
-
-	const bool initialized = context.initialize(compressed_tracks_);
-	ACL_ASSERT(initialized, "Failed to initialize decompression context"); (void)initialized;
-
-	const track_error error = calculate_compression_error(allocator, raw_tracks, context, error_metric, additive_base_tracks);
-	ACL_ASSERT(rtm::scalar_is_finite(error.error), "Returned error is not a finite value"); (void)error;
-	ACL_ASSERT(error.error < regression_error_threshold, "Error too high for bone %u: %f at time %f", error.index, error.error, error.sample_time);
-
-	const uint32_t num_bones = raw_tracks.get_num_tracks();
-	const float clip_duration = raw_tracks.get_duration();
-	const float sample_rate = raw_tracks.get_sample_rate();
-	const uint32_t num_samples = raw_tracks.get_num_samples_per_track();
-
-	debug_track_writer track_writer(allocator, track_type8::qvvf, num_bones);
-
-	// Regression test
-	for (uint32_t sample_index = 0; sample_index < num_samples; ++sample_index)
-	{
-		const float sample_time = rtm::scalar_min(float(sample_index) / sample_rate, clip_duration);
-
-		// We use the nearest sample to accurately measure the loss that happened, if any
-		context.seek(sample_time, sample_rounding_policy::nearest);
-		context.decompress_tracks(track_writer);
-
-		// Validate decompress_track against decompress_tracks
-		for (uint32_t bone_index = 0; bone_index < num_bones; ++bone_index)
-		{
-			const rtm::qvvf transform0 = track_writer.read_qvv(bone_index);
-
-			context.decompress_track(bone_index, track_writer);
-			const rtm::qvvf transform1 = track_writer.read_qvv(bone_index);
-
-			// Rotations can differ a bit due to how we normalize during interpolation
-			ACL_ASSERT(rtm::vector_all_near_equal(rtm::quat_to_vector(transform0.rotation), rtm::quat_to_vector(transform1.rotation), 0.0001F), "Failed to sample bone index: %u", bone_index);
-			ACL_ASSERT(rtm::vector_all_near_equal3(transform0.translation, transform1.translation, 0.0F), "Failed to sample bone index: %u", bone_index);
-			ACL_ASSERT(rtm::vector_all_near_equal3(transform0.scale, transform1.scale, 0.0F), "Failed to sample bone index: %u", bone_index);
-		}
-	}
-}
-
-static void validate_accuracy(iallocator& allocator, const track_array& raw_tracks, const compressed_tracks& tracks, double regression_error_threshold)
-{
-	using namespace acl_impl;
-
-	// Disable floating point exceptions since decompression assumes it
-	scope_disable_fp_exceptions fp_off;
-
-	const float regression_error_thresholdf = static_cast<float>(regression_error_threshold);
-	const rtm::vector4f regression_error_thresholdv = rtm::vector_set(regression_error_thresholdf);
-	(void)regression_error_thresholdf;
-	(void)regression_error_thresholdv;
-
-	const float duration = tracks.get_duration();
-	const float sample_rate = tracks.get_sample_rate();
-	const uint32_t num_tracks = tracks.get_num_tracks();
-	const uint32_t num_samples = tracks.get_num_samples_per_track();
-	const track_type8 track_type = raw_tracks.get_track_type();
-
-	ACL_ASSERT(rtm::scalar_near_equal(duration, raw_tracks.get_duration(), 1.0E-7F), "Duration mismatch");
-	ACL_ASSERT(sample_rate == raw_tracks.get_sample_rate(), "Sample rate mismatch");
-	ACL_ASSERT(num_tracks <= raw_tracks.get_num_tracks(), "Num tracks mismatch");
-	ACL_ASSERT(num_samples == raw_tracks.get_num_samples_per_track(), "Num samples mismatch");
-
-	decompression_context<debug_scalar_decompression_settings> context;
-	context.initialize(tracks);
-
-	debug_track_writer raw_tracks_writer(allocator, track_type, num_tracks);
-	debug_track_writer raw_track_writer(allocator, track_type, num_tracks);
-	debug_track_writer lossy_tracks_writer(allocator, track_type, num_tracks);
-	debug_track_writer lossy_track_writer(allocator, track_type, num_tracks);
-
-	const rtm::vector4f zero = rtm::vector_zero();
-
-	// Regression test
-	for (uint32_t sample_index = 0; sample_index < num_samples; ++sample_index)
-	{
-		const float sample_time = rtm::scalar_min(float(sample_index) / sample_rate, duration);
-
-		// We use the nearest sample to accurately measure the loss that happened, if any
-		raw_tracks.sample_tracks(sample_time, sample_rounding_policy::nearest, raw_tracks_writer);
-
-		context.seek(sample_time, sample_rounding_policy::nearest);
-		context.decompress_tracks(lossy_tracks_writer);
-
-		// Validate decompress_tracks
-		for (uint32_t track_index = 0; track_index < num_tracks; ++track_index)
-		{
-			const track& track_ = raw_tracks[track_index];
-			const uint32_t output_index = track_.get_output_index();
-			if (output_index == k_invalid_track_index)
-				continue;	// Track is being stripped, ignore it
-
-			rtm::vector4f error = zero;
-
-			switch (track_type)
-			{
-			case track_type8::float1f:
-			{
-				const float raw_value = raw_tracks_writer.read_float1(track_index);
-				const float lossy_value = lossy_tracks_writer.read_float1(output_index);
-				error = rtm::vector_set(rtm::scalar_abs(raw_value - lossy_value));
-				break;
-			}
-			case track_type8::float2f:
-			{
-				const rtm::vector4f raw_value = raw_tracks_writer.read_float2(track_index);
-				const rtm::vector4f lossy_value = lossy_tracks_writer.read_float2(output_index);
-				error = rtm::vector_abs(rtm::vector_sub(raw_value, lossy_value));
-				error = rtm::vector_mix<rtm::mix4::x, rtm::mix4::y, rtm::mix4::c, rtm::mix4::d>(error, zero);
-				break;
-			}
-			case track_type8::float3f:
-			{
-				const rtm::vector4f raw_value = raw_tracks_writer.read_float3(track_index);
-				const rtm::vector4f lossy_value = lossy_tracks_writer.read_float3(output_index);
-				error = rtm::vector_abs(rtm::vector_sub(raw_value, lossy_value));
-				error = rtm::vector_mix<rtm::mix4::x, rtm::mix4::y, rtm::mix4::z, rtm::mix4::d>(error, zero);
-				break;
-			}
-			case track_type8::float4f:
-			{
-				const rtm::vector4f raw_value = raw_tracks_writer.read_float4(track_index);
-				const rtm::vector4f lossy_value = lossy_tracks_writer.read_float4(output_index);
-				error = rtm::vector_abs(rtm::vector_sub(raw_value, lossy_value));
-				break;
-			}
-			case track_type8::vector4f:
-			{
-				const rtm::vector4f raw_value = raw_tracks_writer.read_vector4(track_index);
-				const rtm::vector4f lossy_value = lossy_tracks_writer.read_vector4(output_index);
-				error = rtm::vector_abs(rtm::vector_sub(raw_value, lossy_value));
-				break;
-			}
-			default:
-				ACL_ASSERT(false, "Unsupported track type");
-				break;
-			}
-
-			(void)error;
-			ACL_ASSERT(rtm::vector_is_finite(error), "Returned error is not a finite value");
-			ACL_ASSERT(rtm::vector_all_less_than(error, regression_error_thresholdv), "Error too high for track %u at time %f", track_index, sample_time);
-		}
-
-		// Validate decompress_track
-		for (uint32_t track_index = 0; track_index < num_tracks; ++track_index)
-		{
-			const track& track_ = raw_tracks[track_index];
-			const uint32_t output_index = track_.get_output_index();
-			if (output_index == k_invalid_track_index)
-				continue;	// Track is being stripped, ignore it
-
-			// We use the nearest sample to accurately measure the loss that happened, if any
-			raw_tracks.sample_track(track_index, sample_time, sample_rounding_policy::nearest, raw_track_writer);
-			context.decompress_track(output_index, lossy_track_writer);
-
-			switch (track_type)
-			{
-			case track_type8::float1f:
-			{
-				const float raw_value_ = raw_tracks_writer.read_float1(track_index);
-				const float lossy_value_ = lossy_tracks_writer.read_float1(output_index);
-				const float raw_value = raw_track_writer.read_float1(track_index);
-				const float lossy_value = lossy_track_writer.read_float1(output_index);
-				ACL_ASSERT(rtm::scalar_near_equal(raw_value, lossy_value, regression_error_thresholdf), "Error too high for track %u at time %f", track_index, sample_time);
-				ACL_ASSERT(rtm::scalar_near_equal(raw_value_, raw_value, 0.00001F), "Failed to sample track %u at time %f", track_index, sample_time);
-				ACL_ASSERT(rtm::scalar_near_equal(lossy_value_, lossy_value, 0.00001F), "Failed to sample track %u at time %f", track_index, sample_time);
-				break;
-			}
-			case track_type8::float2f:
-			{
-				const rtm::vector4f raw_value_ = raw_tracks_writer.read_float2(track_index);
-				const rtm::vector4f lossy_value_ = lossy_tracks_writer.read_float2(output_index);
-				const rtm::vector4f raw_value = raw_track_writer.read_float2(track_index);
-				const rtm::vector4f lossy_value = lossy_track_writer.read_float2(output_index);
-				ACL_ASSERT(rtm::vector_all_near_equal2(raw_value, lossy_value, regression_error_thresholdf), "Error too high for track %u at time %f", track_index, sample_time);
-				ACL_ASSERT(rtm::vector_all_near_equal2(raw_value_, raw_value, 0.00001F), "Failed to sample track %u at time %f", track_index, sample_time);
-				ACL_ASSERT(rtm::vector_all_near_equal2(lossy_value_, lossy_value, 0.00001F), "Failed to sample track %u at time %f", track_index, sample_time);
-				break;
-			}
-			case track_type8::float3f:
-			{
-				const rtm::vector4f raw_value_ = raw_tracks_writer.read_float3(track_index);
-				const rtm::vector4f lossy_value_ = lossy_tracks_writer.read_float3(output_index);
-				const rtm::vector4f raw_value = raw_track_writer.read_float3(track_index);
-				const rtm::vector4f lossy_value = lossy_track_writer.read_float3(output_index);
-				ACL_ASSERT(rtm::vector_all_near_equal3(raw_value, lossy_value, regression_error_thresholdf), "Error too high for track %u at time %f", track_index, sample_time);
-				ACL_ASSERT(rtm::vector_all_near_equal3(raw_value_, raw_value, 0.00001F), "Failed to sample track %u at time %f", track_index, sample_time);
-				ACL_ASSERT(rtm::vector_all_near_equal3(lossy_value_, lossy_value, 0.00001F), "Failed to sample track %u at time %f", track_index, sample_time);
-				break;
-			}
-			case track_type8::float4f:
-			{
-				const rtm::vector4f raw_value_ = raw_tracks_writer.read_float4(track_index);
-				const rtm::vector4f lossy_value_ = lossy_tracks_writer.read_float4(output_index);
-				const rtm::vector4f raw_value = raw_track_writer.read_float4(track_index);
-				const rtm::vector4f lossy_value = lossy_track_writer.read_float4(output_index);
-				ACL_ASSERT(rtm::vector_all_near_equal(raw_value, lossy_value, regression_error_thresholdf), "Error too high for track %u at time %f", track_index, sample_time);
-				ACL_ASSERT(rtm::vector_all_near_equal(raw_value_, raw_value, 0.00001F), "Failed to sample track %u at time %f", track_index, sample_time);
-				ACL_ASSERT(rtm::vector_all_near_equal(lossy_value_, lossy_value, 0.00001F), "Failed to sample track %u at time %f", track_index, sample_time);
-				break;
-			}
-			case track_type8::vector4f:
-			{
-				const rtm::vector4f raw_value_ = raw_tracks_writer.read_vector4(track_index);
-				const rtm::vector4f lossy_value_ = lossy_tracks_writer.read_vector4(output_index);
-				const rtm::vector4f raw_value = raw_track_writer.read_vector4(track_index);
-				const rtm::vector4f lossy_value = lossy_track_writer.read_vector4(output_index);
-				ACL_ASSERT(rtm::vector_all_near_equal(raw_value, lossy_value, regression_error_thresholdf), "Error too high for track %u at time %f", track_index, sample_time);
-				ACL_ASSERT(rtm::vector_all_near_equal(raw_value_, raw_value, 0.00001F), "Failed to sample track %u at time %f", track_index, sample_time);
-				ACL_ASSERT(rtm::vector_all_near_equal(lossy_value_, lossy_value, 0.00001F), "Failed to sample track %u at time %f", track_index, sample_time);
-				break;
-			}
-			default:
-				ACL_ASSERT(false, "Unsupported track type");
-				break;
-			}
-		}
-	}
-}
-
-static void validate_metadata(const track_array& raw_tracks, const compressed_tracks& tracks)
-{
-	const uint32_t num_tracks = raw_tracks.get_num_tracks();
-
-	// Validate track list name
-	const string& raw_list_name = raw_tracks.get_name();
-	const char* compressed_list_name = tracks.get_name();
-	ACL_ASSERT(raw_list_name == compressed_list_name, "Unexpected track list name");
-
-	// Validate track names
-	for (uint32_t track_index = 0; track_index < num_tracks; ++track_index)
-	{
-		const track& raw_track = raw_tracks[track_index];
-		const uint32_t output_index = raw_track.get_output_index();
-		if (output_index == k_invalid_track_index)
-			continue;	// Stripped
-
-		const string& raw_name = raw_track.get_name();
-		const char* compressed_name = tracks.get_track_name(output_index);
-		ACL_ASSERT(raw_name == compressed_name, "Unexpected track name");
-	}
-
-	if (raw_tracks.get_track_type() == track_type8::qvvf)
-	{
-		// Specific to transform tracks
-		const track_array_qvvf& transform_tracks = track_array_cast<track_array_qvvf>(raw_tracks);
-
-		for (uint32_t track_index = 0; track_index < num_tracks; ++track_index)
-		{
-			const track_qvvf& raw_track = transform_tracks[track_index];
-			const uint32_t output_index = raw_track.get_output_index();
-			if (output_index == k_invalid_track_index)
-				continue;	// Stripped
-
-			const track_desc_transformf& raw_desc = raw_track.get_description();
-			const uint32_t parent_track_index = raw_desc.parent_index;
-			const uint32_t parent_track_output_index = parent_track_index != k_invalid_track_index ? transform_tracks[parent_track_index].get_output_index() : k_invalid_track_index;
-
-			const uint32_t compressed_parent_track_index = tracks.get_parent_track_index(output_index);
-			ACL_ASSERT(parent_track_output_index == compressed_parent_track_index, "Unexpected parent track index");
-
-			track_desc_transformf compressed_desc;
-			const bool compressed_track_desc_found = tracks.get_track_description(output_index, compressed_desc);
-			ACL_ASSERT(compressed_track_desc_found, "Expected track description");
-			ACL_ASSERT(output_index == compressed_desc.output_index, "Unexpected output index");
-			ACL_ASSERT(parent_track_output_index == compressed_desc.parent_index, "Unexpected parent track index");
-			ACL_ASSERT(raw_desc.precision == compressed_desc.precision, "Unexpected precision");
-			ACL_ASSERT(raw_desc.shell_distance == compressed_desc.shell_distance, "Unexpected shell_distance");
-			ACL_ASSERT(raw_desc.constant_rotation_threshold_angle == compressed_desc.constant_rotation_threshold_angle, "Unexpected constant_rotation_threshold_angle");
-			ACL_ASSERT(raw_desc.constant_translation_threshold == compressed_desc.constant_translation_threshold, "Unexpected constant_translation_threshold");
-			ACL_ASSERT(raw_desc.constant_scale_threshold == compressed_desc.constant_scale_threshold, "Unexpected constant_scale_threshold");
-		}
-	}
-	else
-	{
-		// Specific to scalar tracks
-		for (uint32_t track_index = 0; track_index < num_tracks; ++track_index)
-		{
-			const track& raw_track = raw_tracks[track_index];
-			const uint32_t output_index = raw_track.get_output_index();
-			if (output_index == k_invalid_track_index)
-				continue;	// Stripped
-
-			const track_desc_scalarf& raw_desc = raw_track.get_description<track_desc_scalarf>();
-
-			track_desc_scalarf compressed_desc;
-			const bool compressed_track_desc_found = tracks.get_track_description(output_index, compressed_desc);
-			ACL_ASSERT(compressed_track_desc_found, "Expected track description");
-			ACL_ASSERT(output_index == compressed_desc.output_index, "Unexpected output index");
-			ACL_ASSERT(raw_desc.precision == compressed_desc.precision, "Unexpected precision");
-		}
-	}
-}
-
-static void compare_raw_with_compressed(iallocator& allocator, const track_array& raw_tracks, const compressed_tracks& compressed_tracks_)
-{
-	const uint32_t num_tracks = raw_tracks.get_num_tracks();
-
-	uint32_t num_output_tracks = 0;
-	for (uint32_t track_index = 0; track_index < num_tracks; ++track_index)
-	{
-		const track& raw_track = raw_tracks[track_index];
-		const uint32_t output_index = raw_track.get_output_index();
-		if (output_index == k_invalid_track_index)
-			continue;	// Stripped
-
-		num_output_tracks++;
-	}
-
-	ACL_ASSERT(num_output_tracks == compressed_tracks_.get_num_tracks(), "Unexpected num tracks");
-	ACL_ASSERT(raw_tracks.get_num_samples_per_track() == compressed_tracks_.get_num_samples_per_track(), "Unexpected num samples");
-	ACL_ASSERT(raw_tracks.get_sample_rate() == compressed_tracks_.get_sample_rate(), "Unexpected sample rate");
-	ACL_ASSERT(raw_tracks.get_track_type() == compressed_tracks_.get_track_type(), "Unexpected track type");
-	ACL_ASSERT(raw_tracks.get_name() == compressed_tracks_.get_name(), "Unexpected track list name");
-
-	const track_category8 track_category = raw_tracks.get_track_category();
-	for (uint32_t track_index = 0; track_index < num_tracks; ++track_index)
-	{
-		const track& raw_track = raw_tracks[track_index];
-		const uint32_t output_index = raw_track.get_output_index();
-		if (output_index == k_invalid_track_index)
-			continue;	// Stripped
-
-		if (track_category == track_category8::scalarf)
-		{
-			const track_desc_scalarf& raw_desc = raw_track.get_description<track_desc_scalarf>();
-			track_desc_scalarf desc;
-			compressed_tracks_.get_track_description(output_index, desc);
-
-			ACL_ASSERT(raw_desc.precision == desc.precision, "Unexpected precision");
-		}
-		else
-		{
-			const track_desc_transformf& raw_desc = raw_track.get_description<track_desc_transformf>();
-			track_desc_transformf desc;
-			compressed_tracks_.get_track_description(output_index, desc);
-
-			ACL_ASSERT(raw_desc.parent_index == desc.parent_index, "Unexpected parent index");
-			ACL_ASSERT(raw_desc.precision == desc.precision, "Unexpected precision");
-			ACL_ASSERT(raw_desc.shell_distance == desc.shell_distance, "Unexpected shell_distance");
-			ACL_ASSERT(raw_desc.constant_rotation_threshold_angle == desc.constant_rotation_threshold_angle, "Unexpected constant_rotation_threshold_angle");
-			ACL_ASSERT(raw_desc.constant_translation_threshold == desc.constant_translation_threshold, "Unexpected constant_translation_threshold");
-			ACL_ASSERT(raw_desc.constant_scale_threshold == desc.constant_scale_threshold, "Unexpected constant_scale_threshold");
-		}
-	}
-
-	// Disable floating point exceptions since decompression assumes it
-	scope_disable_fp_exceptions fp_off;
-
-	acl::decompression_context<acl_impl::raw_sampling_decompression_settings> context;
-	context.initialize(compressed_tracks_);
-
-	const track_type8 track_type = raw_tracks.get_track_type();
-	acl_impl::debug_track_writer writer(allocator, track_type, num_tracks);
-
-	const uint32_t num_samples = raw_tracks.get_num_samples_per_track();
-	const float sample_rate = raw_tracks.get_sample_rate();
-	const float duration = raw_tracks.get_duration();
-
-	for (uint32_t sample_index = 0; sample_index < num_samples; ++sample_index)
-	{
-		const float sample_time = rtm::scalar_min(float(sample_index) / sample_rate, duration);
-
-		// Round to nearest to land directly on a sample
-		context.seek(sample_time, sample_rounding_policy::nearest);
-
-		context.decompress_tracks(writer);
-
-		for (uint32_t track_index = 0; track_index < num_tracks; ++track_index)
-		{
-			const track& raw_track = raw_tracks[track_index];
-			const uint32_t output_index = raw_track.get_output_index();
-			if (output_index == k_invalid_track_index)
-				continue;	// Track is stripped
-
-			switch (track_type)
-			{
-			case track_type8::float1f:
-			{
-				const float raw_sample = *reinterpret_cast<const float*>(raw_track[sample_index]);
-				const float compressed_sample = writer.read_float1(track_index);
-				ACL_ASSERT(raw_sample == compressed_sample, "Unexpected sample");
-				break;
-			}
-			case track_type8::float2f:
-			{
-				const rtm::vector4f raw_sample = rtm::vector_load2(reinterpret_cast<const rtm::float2f*>(raw_track[sample_index]));
-				const rtm::vector4f compressed_sample = writer.read_float2(track_index);
-				ACL_ASSERT(rtm::vector_all_near_equal2(raw_sample, compressed_sample, 0.0F), "Unexpected sample");
-				break;
-			}
-			case track_type8::float3f:
-			{
-				const rtm::vector4f raw_sample = rtm::vector_load3(reinterpret_cast<const rtm::float3f*>(raw_track[sample_index]));
-				const rtm::vector4f compressed_sample = writer.read_float3(track_index);
-				ACL_ASSERT(rtm::vector_all_near_equal3(raw_sample, compressed_sample, 0.0F), "Unexpected sample");
-				break;
-			}
-			case track_type8::float4f:
-			{
-				const rtm::vector4f raw_sample = rtm::vector_load(reinterpret_cast<const rtm::float4f*>(raw_track[sample_index]));
-				const rtm::vector4f compressed_sample = writer.read_float4(track_index);
-				ACL_ASSERT(rtm::vector_all_near_equal(raw_sample, compressed_sample, 0.0F), "Unexpected sample");
-				break;
-			}
-			case track_type8::vector4f:
-			{
-				const rtm::vector4f raw_sample = *reinterpret_cast<const rtm::vector4f*>(raw_track[sample_index]);
-				const rtm::vector4f compressed_sample = writer.read_vector4(track_index);
-				ACL_ASSERT(rtm::vector_all_near_equal(raw_sample, compressed_sample, 0.0F), "Unexpected sample");
-				break;
-			}
-			case track_type8::qvvf:
-			{
-				const rtm::qvvf raw_sample = *reinterpret_cast<const rtm::qvvf*>(raw_track[sample_index]);
-				const rtm::qvvf compressed_sample = writer.read_qvv(track_index);
-				ACL_ASSERT(rtm::quat_near_equal(raw_sample.rotation, compressed_sample.rotation, 0.0F), "Unexpected sample");
-				ACL_ASSERT(rtm::vector_all_near_equal3(raw_sample.translation, compressed_sample.translation, 0.0F), "Unexpected sample");
-				ACL_ASSERT(rtm::vector_all_near_equal3(raw_sample.scale, compressed_sample.scale, 0.0F), "Unexpected sample");
-				break;
-			}
-			}
-		}
-	}
-}
-
-static void validate_convert(iallocator& allocator, const track_array& raw_tracks)
-{
-	compressed_tracks* compressed_tracks_ = nullptr;
-	error_result result = convert_track_list(allocator, raw_tracks, compressed_tracks_);
-	ACL_ASSERT(result.empty() && compressed_tracks_ != nullptr, "Convert failed");
-
-	compare_raw_with_compressed(allocator, raw_tracks, *compressed_tracks_);
-
-	track_array new_raw_tracks;
-	result = convert_track_list(allocator, *compressed_tracks_, new_raw_tracks);
-	ACL_ASSERT(result.empty() && !new_raw_tracks.is_empty(), "Convert failed");
-
-	compare_raw_with_compressed(allocator, new_raw_tracks, *compressed_tracks_);
-
-	allocator.deallocate(compressed_tracks_, compressed_tracks_->get_size());
-}
-
-static void stream_in_database_tier(database_context<debug_database_settings>& context, debug_database_streamer& streamer, const compressed_database& db, quality_tier tier)
-{
-	const uint32_t num_chunks = db.get_num_chunks(tier);
-
-	bool is_streamed_in = context.is_streamed_in(tier);
-	ACL_ASSERT((num_chunks == 0 && is_streamed_in) || !is_streamed_in, "Tier should not be streamed in");
-	ACL_ASSERT(streamer.get_bulk_data(tier) == nullptr, "Bulk data should not be allocated");
-
-	acl::database_stream_request_result stream_in_result = context.stream_in(tier, 2);
-	const uint8_t* streamer_bulk_data = streamer.get_bulk_data(tier);
-
-	ACL_ASSERT((num_chunks == 0 && stream_in_result == database_stream_request_result::done) || stream_in_result == acl::database_stream_request_result::dispatched, "Failed to stream in tier");
-	ACL_ASSERT(num_chunks == 0 || streamer.get_bulk_data(tier) != nullptr, "Bulk data should be allocated");
-
-	is_streamed_in = context.is_streamed_in(tier);
-	ACL_ASSERT((num_chunks <= 2 && is_streamed_in) || !is_streamed_in, "Failed to stream in tier (first 2 chunks)");
-
-	stream_in_result = context.stream_in(tier);
-
-	ACL_ASSERT((num_chunks <= 2 && stream_in_result == database_stream_request_result::done) || stream_in_result == acl::database_stream_request_result::dispatched, "Failed to stream in tier");
-	ACL_ASSERT(num_chunks == 0 || streamer.get_bulk_data(tier) != nullptr, "Bulk data should be allocated");
-	ACL_ASSERT(streamer.get_bulk_data(tier) == streamer_bulk_data, "Bulk data should not have been reallocated");
-
-	is_streamed_in = context.is_streamed_in(tier);
-	ACL_ASSERT(is_streamed_in, "Failed to stream in tier");
-}
-
-static void stream_out_database_tier(database_context<debug_database_settings>& context, debug_database_streamer& streamer, const compressed_database& db, quality_tier tier)
-{
-	const uint8_t* streamer_bulk_data = streamer.get_bulk_data(tier);
-	const uint32_t num_chunks = db.get_num_chunks(tier);
-
-	const bool is_streamed_in = context.is_streamed_in(tier);
-	ACL_ASSERT(is_streamed_in, "Tier should be streamed in");
-	ACL_ASSERT(num_chunks == 0 || streamer.get_bulk_data(tier) != nullptr, "Bulk data should be allocated");
-
-	acl::database_stream_request_result stream_out_result = context.stream_out(tier, 2);
-
-	ACL_ASSERT((num_chunks == 0 && stream_out_result == database_stream_request_result::done) || stream_out_result == acl::database_stream_request_result::dispatched, "Failed to stream out tier 1");
-	if (num_chunks <= 2)
-	{
-		ACL_ASSERT(streamer.get_bulk_data(tier) == nullptr, "Bulk data should not be allocated");
-	}
-	else
-	{
-		ACL_ASSERT(streamer.get_bulk_data(tier) != nullptr, "Bulk data should be allocated");
-		ACL_ASSERT(streamer.get_bulk_data(tier) == streamer_bulk_data, "Bulk data should not have been reallocated");
-	}
-
-	bool is_streamed_out = !context.is_streamed_in(tier);
-	ACL_ASSERT(num_chunks == 0 || is_streamed_out, "Failed to stream out tier 1 (first 2 chunks)");
-
-	stream_out_result = context.stream_out(tier);
-
-	ACL_ASSERT((num_chunks <= 2 && stream_out_result == database_stream_request_result::done) || stream_out_result == acl::database_stream_request_result::dispatched, "Failed to stream out tier 1");
-	ACL_ASSERT(streamer.get_bulk_data(tier) == nullptr, "Bulk data should not be allocated");
-
-	is_streamed_out = !context.is_streamed_in(tier);
-	ACL_ASSERT(num_chunks == 0 || is_streamed_out, "Failed to stream out tier 1");
-}
-
-static void validate_db_streaming(iallocator& allocator, const track_array_qvvf& raw_tracks, const track_array_qvvf& additive_base_tracks, const itransform_error_metric& error_metric,
-	const track_error& high_quality_tier_error_ref,
-	const compressed_tracks& tracks0, const compressed_tracks& tracks1,
-	const compressed_database& db, const uint8_t* db_bulk_data_medium, const uint8_t* db_bulk_data_low)
-{
-	decompression_context<debug_transform_decompression_settings_with_db> context0;
-	decompression_context<debug_transform_decompression_settings_with_db> context1;
-	database_context<acl::debug_database_settings> db_context;
-	debug_database_streamer db_medium_streamer(allocator, db_bulk_data_medium, db.get_bulk_data_size(quality_tier::medium_importance));
-	debug_database_streamer db_low_streamer(allocator, db_bulk_data_low, db.get_bulk_data_size(quality_tier::lowest_importance));
-	ACL_ASSERT(db_medium_streamer.get_bulk_data(acl::quality_tier::medium_importance) == nullptr, "Bulk data should not be allocated");
-	ACL_ASSERT(db_low_streamer.get_bulk_data(acl::quality_tier::lowest_importance) == nullptr, "Bulk data should not be allocated");
-
-	bool initialized = db_context.initialize(allocator, db, db_medium_streamer, db_low_streamer);
-	initialized = initialized && context0.initialize(tracks0, db_context);
-	initialized = initialized && context1.initialize(tracks1, db_context);
-	ACL_ASSERT(initialized, "Failed to initialize decompression context");
-	ACL_ASSERT(!db_context.is_streamed_in(quality_tier::medium_importance) || db.get_num_chunks(quality_tier::medium_importance) == 0, "Tier shouldn't be streamed in yet");
-	ACL_ASSERT(!db_context.is_streamed_in(quality_tier::lowest_importance) || db.get_num_chunks(quality_tier::lowest_importance) == 0, "Tier shouldn't be streamed in yet");
-
-	// Nothing is streamed in yet, we have low quality
-	const track_error low_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-	ACL_ASSERT(low_quality_tier_error0.error >= high_quality_tier_error_ref.error, "Low quality tier split error should be higher or equal to high quality tier inline");
-	const track_error low_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-	ACL_ASSERT(low_quality_tier_error1.error >= high_quality_tier_error_ref.error, "Low quality tier split error should be higher or equal to high quality tier inline");
-
-	// Stream in our medium importance tier
-	stream_in_database_tier(db_context, db_medium_streamer, db, quality_tier::medium_importance);
-
-	const track_error medium_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-	ACL_ASSERT(medium_quality_tier_error0.error >= high_quality_tier_error_ref.error, "Medium quality tier split error should be higher or equal to high quality tier inline");
-	const track_error medium_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-	ACL_ASSERT(medium_quality_tier_error1.error >= high_quality_tier_error_ref.error, "Medium quality tier split error should be higher or equal to high quality tier inline");
-
-	ACL_ASSERT(low_quality_tier_error0.error >= medium_quality_tier_error0.error, "Low quality tier split error should be higher or equal to medium quality tier split error");
-	ACL_ASSERT(low_quality_tier_error1.error >= medium_quality_tier_error1.error, "Low quality tier split error should be higher or equal to medium quality tier split error");
-
-	// Stream in our low importance tier, restoring the full high quality
-	stream_in_database_tier(db_context, db_low_streamer, db, quality_tier::lowest_importance);
-
-	{
-		const track_error high_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-		ACL_ASSERT(high_quality_tier_error0.error == high_quality_tier_error_ref.error, "High quality tier split error should be equal to high quality tier inline");
-		const track_error high_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-		ACL_ASSERT(high_quality_tier_error1.error == high_quality_tier_error_ref.error, "High quality tier split error should be equal to high quality tier inline");
-	}
-
-	// Stream out our medium importance tier, we'll have mixed quality
-	stream_out_database_tier(db_context, db_medium_streamer, db, quality_tier::medium_importance);
-
-	const track_error mixed_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-	ACL_ASSERT(mixed_quality_tier_error0.error >= high_quality_tier_error_ref.error, "Mixed quality split error should be higher or equal to high quality tier inline");
-	const track_error mixed_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-	ACL_ASSERT(mixed_quality_tier_error1.error >= high_quality_tier_error_ref.error, "Mixed quality split error should be higher or equal to high quality tier inline");
-
-	// Not guaranteed to always be the case due to linear interpolation
-	//ACL_ASSERT(low_quality_tier_error0.error >= mixed_quality_tier_error0.error, "Low quality tier split error should be higher or equal to mixed quality split error");
-	//ACL_ASSERT(low_quality_tier_error1.error >= mixed_quality_tier_error1.error, "Low quality tier split error should be higher or equal to mixed quality split error");
-
-	// Stream in our medium importance tier, restoring the full high quality
-	stream_in_database_tier(db_context, db_medium_streamer, db, quality_tier::medium_importance);
-
-	{
-		const track_error high_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-		ACL_ASSERT(high_quality_tier_error0.error == high_quality_tier_error_ref.error, "High quality tier split error should be equal to high quality tier inline");
-		const track_error high_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-		ACL_ASSERT(high_quality_tier_error1.error == high_quality_tier_error_ref.error, "High quality tier split error should be equal to high quality tier inline");
-	}
-
-	// Stream out our low importance tier, restoring medium quality
-	stream_out_database_tier(db_context, db_low_streamer, db, quality_tier::lowest_importance);
-
-	{
-		const track_error medium_quality_tier_error0_ = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-		ACL_ASSERT(medium_quality_tier_error0_.error == medium_quality_tier_error0.error, "Medium quality should be restored");
-		const track_error medium_quality_tier_error1_ = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-		ACL_ASSERT(medium_quality_tier_error1_.error == medium_quality_tier_error1.error, "Medium quality should be restored");
-	}
-
-	// Stream out our medium importance tier, restoring low quality
-	stream_out_database_tier(db_context, db_medium_streamer, db, quality_tier::medium_importance);
-
-	{
-		const track_error low_quality_tier_error0_ = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-		ACL_ASSERT(low_quality_tier_error0_.error == low_quality_tier_error0.error, "Low quality should be restored");
-		const track_error low_quality_tier_error1_ = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-		ACL_ASSERT(low_quality_tier_error1_.error == low_quality_tier_error1.error, "Low quality should be restored");
-	}
-}
-
-static void validate_db_stripping(iallocator& allocator, const track_array_qvvf& raw_tracks, const track_array_qvvf& additive_base_tracks, const itransform_error_metric& error_metric,
-	const compressed_tracks& tracks0, const compressed_tracks& tracks1,
-	const compressed_database& db, const uint8_t* db_bulk_data_medium, const uint8_t* db_bulk_data_low)
-{
-	compressed_database* db_no_medium = nullptr;
-	compressed_database* db_no_low = nullptr;
-	compressed_database* db_neither0 = nullptr;
-	compressed_database* db_neither1 = nullptr;
-
-	track_error low_quality_tier_error_ref0;
-	track_error low_quality_tier_error_ref1;
-	track_error medium_quality_tier_error_ref0;
-	track_error medium_quality_tier_error_ref1;
-	track_error mixed_quality_tier_error_ref0;
-	track_error mixed_quality_tier_error_ref1;
-
-	// Grab our reference values before we strip anything
-	{
-		decompression_context<debug_transform_decompression_settings_with_db> context0;
-		decompression_context<debug_transform_decompression_settings_with_db> context1;
-		database_context<acl::debug_database_settings> db_context;
-		debug_database_streamer db_medium_streamer(allocator, db_bulk_data_medium, db.get_bulk_data_size(quality_tier::medium_importance));
-		debug_database_streamer db_low_streamer(allocator, db_bulk_data_low, db.get_bulk_data_size(quality_tier::lowest_importance));
-
-		bool initialized = db_context.initialize(allocator, db, db_medium_streamer, db_low_streamer);
-		initialized = initialized && context0.initialize(tracks0, db_context);
-		initialized = initialized && context1.initialize(tracks1, db_context);
-		ACL_ASSERT(initialized, "Failed to initialize decompression context");
-
-		// Nothing is streamed in yet, we have low quality
-		low_quality_tier_error_ref0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-		low_quality_tier_error_ref1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-
-		// Stream in our medium importance tier
-		stream_in_database_tier(db_context, db_medium_streamer, db, quality_tier::medium_importance);
-
-		medium_quality_tier_error_ref0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-		medium_quality_tier_error_ref1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-
-		// Stream in our low importance tier, restoring the full high quality
-		stream_in_database_tier(db_context, db_low_streamer, db, quality_tier::lowest_importance);
-
-		// Stream out our medium importance tier, we'll have mixed quality
-		stream_out_database_tier(db_context, db_medium_streamer, db, quality_tier::medium_importance);
-
-		mixed_quality_tier_error_ref0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-		mixed_quality_tier_error_ref1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-	}
-
-	// Strip medium tier
-	if (db.get_bulk_data_size(quality_tier::medium_importance) != 0)
-	{
-		const error_result result = strip_quality_tier(allocator, db, quality_tier::medium_importance, db_no_medium);
-		ACL_ASSERT(result.empty(), result.c_str());
-
-		decompression_context<debug_transform_decompression_settings_with_db> context0;
-		decompression_context<debug_transform_decompression_settings_with_db> context1;
-		database_context<acl::debug_database_settings> db_context;
-		debug_database_streamer db_medium_streamer(allocator, db_bulk_data_medium, db_no_medium->get_bulk_data_size(quality_tier::medium_importance));
-		debug_database_streamer db_low_streamer(allocator, db_bulk_data_low, db_no_medium->get_bulk_data_size(quality_tier::lowest_importance));
-
-		bool initialized = db_context.initialize(allocator, *db_no_medium, db_medium_streamer, db_low_streamer);
-		initialized = initialized && context0.initialize(tracks0, db_context);
-		initialized = initialized && context1.initialize(tracks1, db_context);
-		ACL_ASSERT(initialized, "Failed to initialize decompression context");
-
-		// Nothing is streamed in yet, we have low quality
-		{
-			const track_error low_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-			ACL_ASSERT(low_quality_tier_error0.error == low_quality_tier_error_ref0.error, "Low quality tier stripped error should be equal to low quality tier whole");
-			const track_error low_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-			ACL_ASSERT(low_quality_tier_error1.error == low_quality_tier_error_ref1.error, "Low quality tier stripped error should be equal to low quality tier whole");
-		}
-
-		// Stream in our medium importance tier
-		stream_in_database_tier(db_context, db_medium_streamer, *db_no_medium, quality_tier::medium_importance);
-
-		{
-			const track_error medium_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-			ACL_ASSERT(medium_quality_tier_error0.error == low_quality_tier_error_ref0.error, "Medium quality tier stripped error should be equal to low quality tier whole");
-			const track_error medium_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-			ACL_ASSERT(medium_quality_tier_error1.error == low_quality_tier_error_ref1.error, "Medium quality tier stripped error should be equal to low quality tier whole");
-		}
-
-		// Stream in our low importance tier, restoring the full high quality
-		stream_in_database_tier(db_context, db_low_streamer, *db_no_medium, quality_tier::lowest_importance);
-
-		{
-			const track_error high_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-			ACL_ASSERT(high_quality_tier_error0.error == mixed_quality_tier_error_ref0.error, "High quality tier stripped error should be equal to mixed quality tier whole");
-			const track_error high_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-			ACL_ASSERT(high_quality_tier_error1.error == mixed_quality_tier_error_ref1.error, "High quality tier stripped error should be equal to mixed quality tier whole");
-		}
-
-		// Stream out our medium importance tier, we'll have mixed quality
-		stream_out_database_tier(db_context, db_medium_streamer, *db_no_medium, quality_tier::medium_importance);
-
-		{
-			const track_error mixed_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-			ACL_ASSERT(mixed_quality_tier_error0.error == mixed_quality_tier_error_ref0.error, "Mixed quality tier stripped error should be equal to mixed quality tier whole");
-			const track_error mixed_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-			ACL_ASSERT(mixed_quality_tier_error1.error == mixed_quality_tier_error_ref1.error, "Mixed quality tier stripped error should be equal to mixed quality tier whole");
-		}
-	}
-
-	// Strip lowest tier
-	if (db.get_bulk_data_size(quality_tier::lowest_importance) != 0)
-	{
-		const error_result result = strip_quality_tier(allocator, db, quality_tier::lowest_importance, db_no_low);
-		ACL_ASSERT(result.empty(), result.c_str());
-
-		decompression_context<debug_transform_decompression_settings_with_db> context0;
-		decompression_context<debug_transform_decompression_settings_with_db> context1;
-		database_context<acl::debug_database_settings> db_context;
-		debug_database_streamer db_medium_streamer(allocator, db_bulk_data_medium, db_no_low->get_bulk_data_size(quality_tier::medium_importance));
-		debug_database_streamer db_low_streamer(allocator, db_bulk_data_low, db_no_low->get_bulk_data_size(quality_tier::lowest_importance));
-
-		bool initialized = db_context.initialize(allocator, *db_no_low, db_medium_streamer, db_low_streamer);
-		initialized = initialized && context0.initialize(tracks0, db_context);
-		initialized = initialized && context1.initialize(tracks1, db_context);
-		ACL_ASSERT(initialized, "Failed to initialize decompression context");
-
-		// Nothing is streamed in yet, we have low quality
-		{
-			const track_error low_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-			ACL_ASSERT(low_quality_tier_error0.error == low_quality_tier_error_ref0.error, "Low quality tier stripped error should be equal to low quality tier whole");
-			const track_error low_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-			ACL_ASSERT(low_quality_tier_error1.error == low_quality_tier_error_ref1.error, "Low quality tier stripped error should be equal to low quality tier whole");
-		}
-
-		// Stream in our medium importance tier
-		stream_in_database_tier(db_context, db_medium_streamer, *db_no_low, quality_tier::medium_importance);
-
-		{
-			const track_error medium_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-			ACL_ASSERT(medium_quality_tier_error0.error == medium_quality_tier_error_ref0.error, "Medium quality tier stripped error should be equal to medium quality tier whole");
-			const track_error medium_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-			ACL_ASSERT(medium_quality_tier_error1.error == medium_quality_tier_error_ref1.error, "Medium quality tier stripped error should be equal to medium quality tier whole");
-		}
-
-		// Stream in our low importance tier, restoring the full high quality
-		stream_in_database_tier(db_context, db_low_streamer, *db_no_low, quality_tier::lowest_importance);
-
-		{
-			const track_error high_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-			ACL_ASSERT(high_quality_tier_error0.error == medium_quality_tier_error_ref0.error, "High quality tier stripped error should be equal to medium quality tier whole");
-			const track_error high_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-			ACL_ASSERT(high_quality_tier_error1.error == medium_quality_tier_error_ref1.error, "High quality tier stripped error should be equal to medium quality tier whole");
-		}
-
-		// Stream out our medium importance tier, we'll have mixed quality
-		stream_out_database_tier(db_context, db_medium_streamer, *db_no_low, quality_tier::medium_importance);
-
-		{
-			const track_error mixed_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-			ACL_ASSERT(mixed_quality_tier_error0.error == low_quality_tier_error_ref0.error, "Mixed quality tier stripped error should be equal to low quality tier whole");
-			const track_error mixed_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-			ACL_ASSERT(mixed_quality_tier_error1.error == low_quality_tier_error_ref1.error, "Mixed quality tier stripped error should be equal to low quality tier whole");
-		}
-	}
-
-	// Strip medium and lowest tiers
-	if (db.get_bulk_data_size(quality_tier::medium_importance) != 0 && db.get_bulk_data_size(quality_tier::lowest_importance) != 0)
-	{
-		ACL_ASSERT(db_no_medium != nullptr, "Expected a valid database");
-		ACL_ASSERT(db_no_low != nullptr, "Expected a valid database");
-
-		error_result result = strip_quality_tier(allocator, *db_no_medium, quality_tier::lowest_importance, db_neither0);
-		ACL_ASSERT(result.empty(), result.c_str());
-		result = strip_quality_tier(allocator, *db_no_low, quality_tier::medium_importance, db_neither1);
-		ACL_ASSERT(result.empty(), result.c_str());
-		ACL_ASSERT(db_neither0->get_hash() == db_neither1->get_hash(), "Stripping order should not matter");
-
-		decompression_context<debug_transform_decompression_settings_with_db> context0;
-		decompression_context<debug_transform_decompression_settings_with_db> context1;
-		database_context<acl::debug_database_settings> db_context;
-		debug_database_streamer db_medium_streamer(allocator, db_bulk_data_medium, db_neither0->get_bulk_data_size(quality_tier::medium_importance));
-		debug_database_streamer db_low_streamer(allocator, db_bulk_data_low, db_neither0->get_bulk_data_size(quality_tier::lowest_importance));
-
-		bool initialized = db_context.initialize(allocator, *db_neither0, db_medium_streamer, db_low_streamer);
-		initialized = initialized && context0.initialize(tracks0, db_context);
-		initialized = initialized && context1.initialize(tracks1, db_context);
-		ACL_ASSERT(initialized, "Failed to initialize decompression context");
-
-		// Nothing is streamed in yet, we have low quality
-		{
-			const track_error low_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-			ACL_ASSERT(low_quality_tier_error0.error == low_quality_tier_error_ref0.error, "Low quality tier stripped error should be equal to low quality tier whole");
-			const track_error low_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-			ACL_ASSERT(low_quality_tier_error1.error == low_quality_tier_error_ref1.error, "Low quality tier stripped error should be equal to low quality tier whole");
-		}
-
-		// Stream in our medium importance tier
-		stream_in_database_tier(db_context, db_medium_streamer, *db_neither0, quality_tier::medium_importance);
-
-		{
-			const track_error medium_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-			ACL_ASSERT(medium_quality_tier_error0.error == low_quality_tier_error_ref0.error, "Medium quality tier stripped error should be equal to low quality tier whole");
-			const track_error medium_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-			ACL_ASSERT(medium_quality_tier_error1.error == low_quality_tier_error_ref1.error, "Medium quality tier stripped error should be equal to low quality tier whole");
-		}
-
-		// Stream in our low importance tier, restoring the full high quality
-		stream_in_database_tier(db_context, db_low_streamer, *db_neither0, quality_tier::lowest_importance);
-
-		{
-			const track_error high_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-			ACL_ASSERT(high_quality_tier_error0.error == low_quality_tier_error_ref0.error, "High quality tier stripped error should be equal to low quality tier whole");
-			const track_error high_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-			ACL_ASSERT(high_quality_tier_error1.error == low_quality_tier_error_ref1.error, "High quality tier stripped error should be equal to low quality tier whole");
-		}
-
-		// Stream out our medium importance tier, we'll have mixed quality
-		stream_out_database_tier(db_context, db_medium_streamer, *db_neither0, quality_tier::medium_importance);
-
-		{
-			const track_error mixed_quality_tier_error0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-			ACL_ASSERT(mixed_quality_tier_error0.error == low_quality_tier_error_ref0.error, "Mixed quality tier stripped error should be equal to low quality tier whole");
-			const track_error mixed_quality_tier_error1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-			ACL_ASSERT(mixed_quality_tier_error1.error == low_quality_tier_error_ref1.error, "Mixed quality tier stripped error should be equal to low quality tier whole");
-		}
-	}
-
-	if (db_no_medium != nullptr)
-		allocator.deallocate(db_no_medium, db_no_medium->get_size());
-	if (db_no_low != nullptr)
-		allocator.deallocate(db_no_low, db_no_low->get_size());
-	if (db_neither0 != nullptr)
-		allocator.deallocate(db_neither0, db_neither0->get_size());
-	if (db_neither1 != nullptr)
-		allocator.deallocate(db_neither1, db_neither1->get_size());
-}
-
-static void validate_db(iallocator& allocator, const track_array_qvvf& raw_tracks, const track_array_qvvf& additive_base_tracks,
+// We extern our regression test functions for simplicity
+extern void validate_accuracy(iallocator& allocator, const track_array_qvvf& raw_tracks, const track_array_qvvf& additive_base_tracks, const itransform_error_metric& error_metric, const compressed_tracks& compressed_tracks_, double regression_error_threshold);
+extern void validate_accuracy(iallocator& allocator, const track_array& raw_tracks, const compressed_tracks& tracks, double regression_error_threshold);
+extern void validate_metadata(const track_array& raw_tracks, const compressed_tracks& tracks);
+extern void validate_convert(iallocator& allocator, const track_array& raw_tracks);
+extern void validate_db(iallocator& allocator, const track_array_qvvf& raw_tracks, const track_array_qvvf& additive_base_tracks,
 	const compression_database_settings& settings, const itransform_error_metric& error_metric,
-	const compressed_tracks& compressed_tracks0, const compressed_tracks& compressed_tracks1)
-{
-	using namespace acl_impl;
-
-	// Disable floating point exceptions since decompression assumes it
-	scope_disable_fp_exceptions fp_off;
-
-	// Build our databases
-	const compressed_tracks* input_tracks[2] = { &compressed_tracks0, &compressed_tracks1 };
-
-	compressed_tracks* db_tracks0[1] = { nullptr };
-	compressed_tracks* db_tracks1[1] = { nullptr };
-	compressed_tracks* db_tracks01[2] = { nullptr, nullptr };
-	compressed_database* db0 = nullptr;
-	compressed_database* db1 = nullptr;
-	compressed_database* db01 = nullptr;
-
-	{
-		error_result db_result = build_database(allocator, settings, &input_tracks[0], 1, db_tracks0, db0);
-		ACL_ASSERT(db_result.empty(), db_result.c_str());
-
-		db_result = build_database(allocator, settings, &input_tracks[1], 1, db_tracks1, db1);
-		ACL_ASSERT(db_result.empty(), db_result.c_str());
-
-		db_result = build_database(allocator, settings, &input_tracks[0], 2, db_tracks01, db01);
-		ACL_ASSERT(db_result.empty(), db_result.c_str());
-
-		ACL_ASSERT(db0->contains(*db_tracks0[0]), "Database should contain our clip");
-		ACL_ASSERT(db1->contains(*db_tracks1[0]), "Database should contain our clip");
-		ACL_ASSERT(db01->contains(*db_tracks01[0]), "Database should contain our clip");
-		ACL_ASSERT(db01->contains(*db_tracks01[1]), "Database should contain our clip");
-	}
-
-	// Reference error with the bulk data inline and everything loaded
-	track_error high_quality_tier_error_ref;
-	{
-		acl::decompression_context<debug_transform_decompression_settings_with_db> context;
-
-		const bool initialized = context.initialize(compressed_tracks0);
-		ACL_ASSERT(initialized, "Failed to initialize decompression context");
-
-		high_quality_tier_error_ref = calculate_compression_error(allocator, raw_tracks, context, error_metric, additive_base_tracks);
-	}
-
-	// Make sure the databases agree with our reference
-	{
-		acl::decompression_context<debug_transform_decompression_settings_with_db> context;
-		acl::database_context<acl::debug_database_settings> db_context;
-
-		bool initialized = db_context.initialize(allocator, *db0);
-		initialized = initialized && context.initialize(*db_tracks0[0], db_context);
-		ACL_ASSERT(initialized, "Failed to initialize decompression context");
-
-		const track_error error_tier0 = calculate_compression_error(allocator, raw_tracks, context, error_metric, additive_base_tracks);
-		ACL_ASSERT(error_tier0.error == high_quality_tier_error_ref.error, "Database 0 should have the same error");
-	}
-
-	{
-		acl::decompression_context<debug_transform_decompression_settings_with_db> context;
-		acl::database_context<acl::debug_database_settings> db_context;
-
-		bool initialized = db_context.initialize(allocator, *db1);
-		initialized = initialized && context.initialize(*db_tracks1[0], db_context);
-		ACL_ASSERT(initialized, "Failed to initialize decompression context");
-
-		const track_error error_tier1 = calculate_compression_error(allocator, raw_tracks, context, error_metric, additive_base_tracks);
-		ACL_ASSERT(error_tier1.error == high_quality_tier_error_ref.error, "Database 1 should have the same error");
-	}
-
-	{
-		acl::decompression_context<debug_transform_decompression_settings_with_db> context0;
-		acl::decompression_context<debug_transform_decompression_settings_with_db> context1;
-		acl::database_context<acl::debug_database_settings> db_context;
-
-		bool initialized = db_context.initialize(allocator, *db01);
-		initialized = initialized && context0.initialize(*db_tracks01[0], db_context);
-		initialized = initialized && context1.initialize(*db_tracks01[1], db_context);
-		ACL_ASSERT(initialized, "Failed to initialize decompression context");
-
-		const track_error error_tier0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-		ACL_ASSERT(error_tier0.error == high_quality_tier_error_ref.error, "Database 01 should have the same error");
-
-		const track_error error_tier1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-		ACL_ASSERT(error_tier1.error == high_quality_tier_error_ref.error, "Database 01 should have the same error");
-	}
-
-	// Split the database bulk data out
-	compressed_database* split_db = nullptr;
-	uint8_t* split_db_bulk_data_medium = nullptr;
-	uint8_t* split_db_bulk_data_low = nullptr;
-	const error_result split_result = split_compressed_database_bulk_data(allocator, *db01, split_db, split_db_bulk_data_medium, split_db_bulk_data_low);
-	ACL_ASSERT(split_result.empty(), "Failed to split database");
-	ACL_ASSERT(split_db->is_valid(true).empty(), "Failed to split database");
-
-	ACL_ASSERT(split_db->contains(*db_tracks01[0]), "Database should contain our clip");
-	ACL_ASSERT(split_db->contains(*db_tracks01[1]), "Database should contain our clip");
-
-	// Measure the tier error through simulated streaming
-	validate_db_streaming(allocator, raw_tracks, additive_base_tracks, error_metric, high_quality_tier_error_ref, *db_tracks01[0], *db_tracks01[1], *split_db, split_db_bulk_data_medium, split_db_bulk_data_low);
-
-	// Measure the tier error when stripping
-	validate_db_stripping(allocator, raw_tracks, additive_base_tracks, error_metric, *db_tracks01[0], *db_tracks01[1], *split_db, split_db_bulk_data_medium, split_db_bulk_data_low);
-
-	// Duplicate our clips so we can modify them
-	compressed_tracks* compressed_tracks_copy0 = safe_ptr_cast<compressed_tracks>(allocate_type_array_aligned<uint8_t>(allocator, db_tracks0[0]->get_size(), alignof(compressed_tracks)));
-	compressed_tracks* compressed_tracks_copy1 = safe_ptr_cast<compressed_tracks>(allocate_type_array_aligned<uint8_t>(allocator, db_tracks1[0]->get_size(), alignof(compressed_tracks)));
-	std::memcpy(reinterpret_cast<uint8_t*>(compressed_tracks_copy0), db_tracks0[0], db_tracks0[0]->get_size());
-	std::memcpy(reinterpret_cast<uint8_t*>(compressed_tracks_copy1), db_tracks1[0], db_tracks1[0]->get_size());
-
-	// Merge our everything into a new database
-	database_merge_mapping mappings[2];
-	mappings[0].tracks = compressed_tracks_copy0;
-	mappings[0].database = db0;
-	mappings[1].tracks = compressed_tracks_copy1;
-	mappings[1].database = db1;
-
-	compressed_database* merged_db = nullptr;
-	const error_result merge_result = merge_compressed_databases(allocator, settings, &mappings[0], 2, merged_db);
-	ACL_ASSERT(merge_result.empty(), "Failed to merge databases");
-	ACL_ASSERT(merged_db->is_valid(true).empty(), "Failed to merge database");
-
-	ACL_ASSERT(merged_db->contains(*compressed_tracks_copy0), "New database should contain our clip");
-	ACL_ASSERT(merged_db->contains(*compressed_tracks_copy1), "New database should contain our clip");
-
-	{
-		acl::decompression_context<debug_transform_decompression_settings_with_db> context0;
-		acl::decompression_context<debug_transform_decompression_settings_with_db> context1;
-		acl::database_context<acl::debug_database_settings> db_context;
-
-		bool initialized = db_context.initialize(allocator, *merged_db);
-		initialized = initialized && context0.initialize(*compressed_tracks_copy0, db_context);
-		initialized = initialized && context1.initialize(*compressed_tracks_copy1, db_context);
-		ACL_ASSERT(initialized, "Failed to initialize decompression context");
-
-		const track_error error_tier1_ref_merged0 = calculate_compression_error(allocator, raw_tracks, context0, error_metric, additive_base_tracks);
-		ACL_ASSERT(high_quality_tier_error_ref.error == error_tier1_ref_merged0.error, "Reference error should be equal to merged error");
-		const track_error error_tier1_ref_merged1 = calculate_compression_error(allocator, raw_tracks, context1, error_metric, additive_base_tracks);
-		ACL_ASSERT(high_quality_tier_error_ref.error == error_tier1_ref_merged1.error, "Reference error should be equal to merged error");
-	}
-
-	// Split the database bulk data out
-	compressed_database* split_merged_db = nullptr;
-	uint8_t* split_merged_db_bulk_data_medium = nullptr;
-	uint8_t* split_merged_db_bulk_data_low = nullptr;
-	const error_result split_merge_result = split_compressed_database_bulk_data(allocator, *merged_db, split_merged_db, split_merged_db_bulk_data_medium, split_merged_db_bulk_data_low);
-	ACL_ASSERT(split_merge_result.empty(), "Failed to split merged database");
-	ACL_ASSERT(split_merged_db->is_valid(true).empty(), "Failed to split merged database");
-
-	ACL_ASSERT(split_merged_db->contains(*compressed_tracks_copy0), "New database should contain our clip");
-	ACL_ASSERT(split_merged_db->contains(*compressed_tracks_copy1), "New database should contain our clip");
-
-	// Measure the tier error through simulated streaming
-	validate_db_streaming(allocator, raw_tracks, additive_base_tracks, error_metric, high_quality_tier_error_ref, *compressed_tracks_copy0, *compressed_tracks_copy1, *split_merged_db, split_merged_db_bulk_data_medium, split_merged_db_bulk_data_low);
-
-	// Measure the tier error when stripping
-	validate_db_stripping(allocator, raw_tracks, additive_base_tracks, error_metric, *compressed_tracks_copy0, *compressed_tracks_copy1, *split_merged_db, split_merged_db_bulk_data_medium, split_merged_db_bulk_data_low);
-
-	// Free our memory
-	allocator.deallocate(split_db_bulk_data_medium, split_db->get_bulk_data_size(quality_tier::medium_importance));
-	allocator.deallocate(split_db_bulk_data_low, split_db->get_bulk_data_size(quality_tier::lowest_importance));
-	allocator.deallocate(split_db, split_db->get_size());
-	allocator.deallocate(compressed_tracks_copy0, compressed_tracks_copy0->get_size());
-	allocator.deallocate(compressed_tracks_copy1, compressed_tracks_copy1->get_size());
-	allocator.deallocate(split_merged_db_bulk_data_medium, split_merged_db->get_bulk_data_size(quality_tier::medium_importance));
-	allocator.deallocate(split_merged_db_bulk_data_low, split_merged_db->get_bulk_data_size(quality_tier::lowest_importance));
-	allocator.deallocate(split_merged_db, split_merged_db->get_size());
-	allocator.deallocate(merged_db, merged_db->get_size());
-	allocator.deallocate(db_tracks0[0], db_tracks0[0]->get_size());
-	allocator.deallocate(db_tracks1[0], db_tracks1[0]->get_size());
-	allocator.deallocate(db_tracks01[0], db_tracks01[0]->get_size());
-	allocator.deallocate(db_tracks01[1], db_tracks01[1]->get_size());
-	allocator.deallocate(db0, db0->get_size());
-	allocator.deallocate(db1, db1->get_size());
-	allocator.deallocate(db01, db01->get_size());
-}
+	const compressed_tracks& compressed_tracks0, const compressed_tracks& compressed_tracks1);
 #endif
 
 static void try_algorithm(const Options& options, iallocator& allocator, track_array_qvvf& transform_tracks,
@@ -1509,7 +491,7 @@ static void try_algorithm(const Options& options, iallocator& allocator, track_a
 			// Disable floating point exceptions since decompression assumes it
 			scope_disable_fp_exceptions fp_off;
 
-			acl::decompression_context<debug_transform_decompression_settings_with_db> context;
+			decompression_context<debug_transform_decompression_settings> context;
 
 			const bool initialized = context.initialize(*compressed_tracks_);
 			ACL_ASSERT(initialized, "Failed to initialize decompression context"); (void)initialized;
